server/requirements.txt
numpy==1.24.1
opencv-python==4.7.0.68
pafy==0.5.5
python-dotenv==0.21.1
youtube-dl==2020.12.2
